upate for move; dropping asf-tools references
