--- .github/CODEOWNERS
@@ -1,2 +1,0 @@
-# These owners will be requested for review when someone opens a pull request.
-*       @ASFHyP3/plugins
