--- .github/PULL_REQUEST_TEMPLATE/release.md
@@ -1,37 +1,0 @@
-## Description of this release
-
-<!--
-Please describe the release here, including a brief overview of the changes in this release
--->
-
-<!--
-If applicable, indicate any upstream packages/projects this is relevant too, and the associated issues
-or pull requests
--->
-
-### Developer checklist
-
-- [ ] Assigned a reviewer
-  <!-- NOTE: Pull requests should only be opened for merges to protected branches (required) and any
-   changes which you'd like reviewed. Do not open a pull request to update a feature or personal
-   branch -- simply merge with `git`.
-   -->
-- [ ] Indicated the level of changes to this package by affixing one of these labels:
-  * major -- Major changes to the API that may break current workflows
-  * minor -- Minor changes to the API that do not break current workflows
-  * patch -- Patches and bugfixes for the current version that do not break current workflows
-  * bumpless -- Changes to documentation, CI/CD pipelines, etc. that don't affect the software's version
-
-- [ ] (If applicable) Updated the dependencies and indicated any downstream changes that are required
-
-- [ ] Updated `CHANGELOG.md`
-- [ ] Added/updated documentation for these changes
-- [ ] Added/updated tests for these changes
-
-### Reviewer checklist
-
-- [ ] Have all dependencies been updated?
-- [ ] Is the level of changes labeled appropriately?
-- [ ] Are all the changes described appropriately in `CHANGELOG.md`?
-- [ ] Has the documentation been adequately updated?
-- [ ] Are the tests adequate?
