--- .github/pull_request_template.md
@@ -1,8 +1,0 @@
-<!--
-If this is a pull request for a new release, please use the release template:
-   https://github.com/ASFHyP3/asf-tools/compare/main...develop?template=release.md
- 
-NOTE: Pull requests should only be opened for merges to protected branches (required) and any 
-changes which you'd like reviewed. Do not open a pull request to update a feature or personal
-branch -- simply merge with `git`.
--->
