+++ Dockerfile
@@ -2,14 +2,13 @@ FROM condaforge/mambaforge:latest
 
 # For opencontainers label definitions, see:
 #    https://github.com/opencontainers/image-spec/blob/master/annotations.md
-LABEL org.opencontainers.image.title="ASF Tools"
-LABEL org.opencontainers.image.description="Plugin for tools utilized in SAR processing"
+LABEL org.opencontainers.image.title="HydroSAR"
+LABEL org.opencontainers.image.description="HyP3 Plugin for monitoring of hydrological hazards"
 LABEL org.opencontainers.image.vendor="Alaska Satellite Facility"
-LABEL org.opencontainers.image.authors="ASF APD/Tools Team <uaf-asf-apd@alaska.edu>"
 LABEL org.opencontainers.image.licenses="BSD-3-Clause"
-LABEL org.opencontainers.image.url="https://github.com/ASFHyP3/asf-tools"
-LABEL org.opencontainers.image.source="https://github.com/ASFHyP3/asf-tools"
-LABEL org.opencontainers.image.documentation="https://hyp3-docs.asf.alaska.edu"
+LABEL org.opencontainers.image.url="https://github.com/fjmeyer/HydroSAR"
+LABEL org.opencontainers.image.source="https://github.com/fjmeyer/HydroSAR"
+LABEL org.opencontainers.image.documentation="https://github.com/fjmeyer/HydroSAR"
 
 ARG DEBIAN_FRONTEND=noninteractive
 ENV PYTHONDONTWRITEBYTECODE=true
@@ -30,16 +29,16 @@ USER ${CONDA_UID}
 SHELL ["/bin/bash", "-l", "-c"]
 WORKDIR /home/conda/
 
-COPY --chown=${CONDA_UID}:${CONDA_GID} . /asf-tools/
+COPY --chown=${CONDA_UID}:${CONDA_GID} . /hydrosar/
 
-RUN mamba env create -f /asf-tools/environment.yml && \
+RUN mamba env create -f /hydrosar/environment.yml && \
     conda clean -afy && \
-    conda activate asf-tools && \
-    sed -i 's/conda activate base/conda activate asf-tools/g' /home/conda/.profile && \
-    python -m pip install --no-cache-dir /asf-tools
+    conda activate hydrosar && \
+    sed -i 's/conda activate base/conda activate hydrosar/g' /home/conda/.profile && \
+    python -m pip install --no-cache-dir /hydrosar
 
-# attempt to speed up initial asf_tools imports by caching *something*...
-RUN python -m asf_tools --help
+# attempt to speed up initial hydrosar imports by caching *something*...
+RUN python -m hydrosar --help
 
-ENTRYPOINT ["/asf-tools/src/asf_tools/etc/entrypoint.sh"]
+ENTRYPOINT ["/hydrosar/src/hydrosar/etc/entrypoint.sh"]
 CMD ["-h"]
