+++ MANIFEST.in
@@ -1,4 +1,1 @@
 global-exclude *.py[cod] __pycache__ *.so
-
-prune ArcGIS-toolbox
-prune prototype
