--- README-asf-tools.md
@@ -1,46 +1,0 @@
-ASF Tools
-=========
-**_A collection of tools for working with higher-level SAR products_**
------
-
-[![DOI](https://zenodo.org/badge/295506894.svg)](https://zenodo.org/badge/latestdoi/295506894)
-
-## ArcGIS Toolbox
-The [ArcGIS Python Toolbox](ArcGIS-toolbox) can be used with either ArcGIS Desktop or 
-ArcGIS Pro, and contains tools that perform geoprocessing tasks useful for working
-with Synthetic Aperture Radar (SAR) data. The tools were designed to be used with
-Radiometric Terrain Corrected (RTC) SAR datasets, such as those enerated by
-[HyP3](https://hyp3-docs.asf.alaska.edu/), but several of the tools have the
-potential to be used with a variety of rasters, including non-SAR datasets.
-
-More information and a download link is also provided on
-[ASF's website](https://asf.alaska.edu/how-to/data-tools/gis-tools/)
-
-The Toolbox is distributed as a zipped archive including the `.pyt` Toolbox script
-and associated `.xml` files. There is an XML file for the toolbox itself and one
-for each of the tools it contains. These XML files contain the metadata displayed
-in the item descriptions and tool help windows, and must be kept in the same directory 
-as the Python Toolbox (`.pyt`) file, or the information they contain will no longer
-be accessible to ArcGIS.
-
-## ASF Tools for Python
-
-[asf_tools](src/asf_tools) is a Python package for working with Synthetic Aperture Radar (SAR) data.
-It was designed for working with datasets generated by
-[HyP3](https://hyp3-docs.asf.alaska.edu/), but several of the tools have the
-potential to be used with a variety of rasters, including non-SAR datasets.
-
-`asf_tools` is distributed on [PyPI](https://pypi.org/) and 
-[conda-forge](https://conda-forge.org/docs/user/introduction.html). See our
-[documentation](https://hyp3-docs.asf.alaska.edu/tools/asf_tools/) for installation
-and usage.
-
-## Contact Us
-
-Want to talk about ASF Tools? We would love to hear from you!
-
-Found a bug? Want to request a feature?
-[open an issue](https://github.com/ASFHyP3/asf_tools/issues/new)
-
-General questions? Suggestions? Or just want to talk to the team?
-[chat with us on gitter](https://gitter.im/ASFHyP3/community)
