+++ src/hydrosar/__init__.py
@@ -1,0 +1,9 @@
+"""Algorithms for the monitoring of hydrological hazards using data from Synthetic Aperture Radar (SAR) sensors"""
+
+from importlib.metadata import version
+
+__version__ = version(__name__)
+
+__all__ = [
+    '__version__',
+]
