+++ src/hydrosar/hand/__init__.py
@@ -1,5 +1,5 @@
-from asf_tools.hydrosar.hand.calculate import calculate_hand_for_basins, make_copernicus_hand
-from asf_tools.hydrosar.hand.prepare import prepare_hand_vrt
+from hydrosar.hand.calculate import calculate_hand_for_basins, make_copernicus_hand
+from hydrosar.hand.prepare import prepare_hand_vrt
 
 __all__ = [
     'calculate_hand_for_basins',
