+++ src/hydrosar/hand/prepare.py
@@ -3,13 +3,12 @@ from pathlib import Path
 from tempfile import NamedTemporaryFile
 from typing import Union
 
+from asf_tools import vector
+from asf_tools.util import GDALConfigManager, get_epsg_code
 from osgeo import gdal, ogr
 from rasterio.enums import Resampling
 from shapely.geometry import shape
 from shapely.geometry.base import BaseGeometry
-
-from asf_tools import vector
-from asf_tools.util import GDALConfigManager, get_epsg_code
 
 HAND_GEOJSON = '/vsicurl/https://glo-30-hand.s3.amazonaws.com/v1/2021/glo-30-hand.geojson'
