+++ src/hydrosar/water_map.py
@@ -15,15 +15,15 @@ from typing import Literal, Optional, Tuple, Union
 
 import numpy as np
 import skfuzzy as fuzz
-from osgeo import gdal
-from skimage import measure
-
 from asf_tools.aws import get_path_to_s3_file, upload_file_to_s3
-from asf_tools.hydrosar.hand.prepare import prepare_hand_for_raster
-from asf_tools.hydrosar.threshold import expectation_maximization_threshold as em_threshold
 from asf_tools.raster import read_as_masked_array, write_cog
 from asf_tools.tile import tile_array, untile_array
 from asf_tools.util import get_epsg_code
+from osgeo import gdal
+from skimage import measure
+
+from hydrosar.hand.prepare import prepare_hand_for_raster
+from hydrosar.threshold import expectation_maximization_threshold as em_threshold
 
 log = logging.getLogger(__name__)
