+++ tests/test_hand.py
@@ -1,12 +1,12 @@
 import json
 
+import numpy as np
 import pytest
+from asf_tools import vector
 from osgeo import gdal, ogr
 from osgeo_utils.gdalcompare import find_diff
-import numpy as np
 
-from asf_tools import vector
-from asf_tools.hydrosar import hand
+from hydrosar import hand
 
 HAND_BASINS = '/vsicurl/https://hyp3-testing.s3-us-west-2.amazonaws.com/' \
               'asf-tools/S1A_IW_20230228T120437_DVR_RTC30/hand/hybas_af_lev12_v1c_firstpoly.geojson'
